Initial import of the data
This comes from https://maps.amsterdam.nl/open_geodata/?k=240

The terms are here: https://maps.amsterdam.nl/open_geodata/terms.php?LANG=en

License
The City of Amsterdam (hereinafter the City) now grants you a license to use and re-use the downloaded dataset for any lawful purpose. You may use the dataset both for non-commercial and commercial purposes.
You now acknowledge that this license does not give you a copyright or other proprietary interest in the dataset.
Warranty
The data in the dataset are initially generated by the City for its own applications. If the municipal process, in which the dataset is used or of which the dataset is a product, expires, the dataset will also expire. Therefore you can not claim to make the dataset available and the City does not give you warranties or commitments on the dataset, none of any kind. By this we mean for example, but without limiting the total scope:

    While reasonable efforts have been made in preparing the dataset for use by you, the City cannot give any warranties as to the completeness, currency, or accuracy of the dataset.
    The City may at any time and from time to time change the dataset, no longer make it available or otherwise suspend access to it, without a transitional period being observed and without personal notice.
    The City can not, at your request, add extra data to the dataset, update the dataset or distribute the dataset in another format or coördinate system.
    Historical datasets will not be retained, unless this is indicated for the purpose of the dataset.

Citation
When re-using the dataset, you are not required to mention the City of Amsterdam as the source holder, but it is appreciated if possible. You may not imply that the City is endorsing the purpose of your derived work.
You may not use any official trade-mark, emblem or logo of the City, or any of its other references without the City's prior written consent.
In case you distribute the dataset as a dataset, with fee or for free, in its original or in modified form, you are required to mention that the original dataset is freely available from the City for everyone, under these Terms of Use, which also apply to the distributed dataset.
Your and our liability
The City excludes any liability for any direct or indirect damage whatsoever, caused in any way by or arising from any use of the dataset, including - but not only - actions of a user of the dataset which would have been prompted by available data.